+++ src/main/java/com/db/bexlibrary/BexLibrary/controllers/LoanController.java
@@ -1,8 +1,9 @@
 package com.db.bexlibrary.BexLibrary.controllers;
 
 import com.db.bexlibrary.BexLibrary.entities.Loan;
-import com.db.bexlibrary.BexLibrary.entities.LoanPOJO;
+import com.db.bexlibrary.BexLibrary.pojos.LoanPOJO;
 import com.db.bexlibrary.BexLibrary.javamail.MailSender;
+import com.db.bexlibrary.BexLibrary.pojos.ReturnedBookPOJO;
 import com.db.bexlibrary.BexLibrary.repositories.BookRepo;
 import com.db.bexlibrary.BexLibrary.repositories.LoanRepo;
 import com.db.bexlibrary.BexLibrary.repositories.UserRepo;
@@ -48,4 +49,9 @@ public class LoanController {
     return new ResponseEntity<List<Loan>>(loanService.getAllLoans(), HttpStatus.OK);
   }
 
+  @RequestMapping(value = "/return", method = RequestMethod.POST)
+  public  ResponseEntity<?> returnBook(@RequestBody ReturnedBookPOJO pojo){
+    loanService.returnBookMethod(pojo.getBookId(),pojo.getLoanId());
+    return new ResponseEntity<>(HttpStatus.OK);
+  }
 }
