--- src/main/java/com/db/bexlibrary/BexLibrary/entities/UserPOJO.java
@@ -1,12 +1,0 @@
-package com.db.bexlibrary.BexLibrary.entities;
-
-import lombok.Data;
-
-@Data
-public class UserPOJO {
-
-  private String userEmail;
-
-  private String password;
-
-}
