+++ src/main/java/com/db/bexlibrary/BexLibrary/pojos/ReturnedBookPOJO.java
@@ -1,0 +1,12 @@
+package com.db.bexlibrary.BexLibrary.pojos;
+
+import lombok.Data;
+
+@Data
+public class ReturnedBookPOJO {
+
+  private Long bookId;
+
+  private Long loanId;
+
+}
