+++ src/main/resources/application.properties
@@ -3,7 +3,7 @@ spring.datasource.username=masterBexlibrary
 spring.datasource.password=b3xlibrary!
 
 #spring.datasource.driver-class-name=
-spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.hibernate.ddl-auto=update
 spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
